+++ tests/src/vec2_access_tests.ads
@@ -1,0 +1,5 @@
+with AUnit.Test_Suites;
+
+package Vec2_Access_Tests is
+   function Suite return AUnit.Test_Suites.Access_Test_Suite;
+end Vec2_Access_Tests;